requirements.txt
pyyaml==3.12
psycopg2==2.6.2
prometheus_client==0.0.16
